+++ tests/config.test.ts
@@ -143,30 +143,23 @@ Deno.test("pages configuration", () => {
 
   equals(formats.size, 10);
   equals(formats.has(".tmpl.json"), true);
-  assert(formats.get(".tmpl.json")?.removeExtension);
-  assert(formats.get(".tmpl.json")?.page);
+  equals(formats.get(".tmpl.json")?.page, "html");
   equals(formats.has(".tmpl.js"), true);
-  assert(formats.get(".tmpl.js")?.removeExtension);
-  assert(formats.get(".tmpl.js")?.page);
+  equals(formats.get(".tmpl.js")?.page, "html");
   assert(formats.get(".tmpl.js")?.engine);
   equals(formats.has(".tmpl.ts"), true);
-  assert(formats.get(".tmpl.ts")?.removeExtension);
-  assert(formats.get(".tmpl.ts")?.page);
+  equals(formats.get(".tmpl.ts")?.page, "html");
   assert(formats.get(".tmpl.ts")?.engine);
   equals(formats.has(".md"), true);
-  assert(formats.get(".md")?.removeExtension);
-  assert(formats.get(".md")?.page);
+  equals(formats.get(".md")?.page, "html");
   assert(formats.get(".md")?.engine);
   equals(formats.has(".njk"), true);
-  assert(formats.get(".njk")?.removeExtension);
-  assert(formats.get(".njk")?.page);
+  equals(formats.get(".njk")?.page, "html");
   assert(formats.get(".njk")?.engine);
   equals(formats.has(".yaml"), true);
-  assert(formats.get(".yaml")?.removeExtension);
-  assert(formats.get(".yaml")?.page);
+  equals(formats.get(".yaml")?.page, "html");
   equals(formats.has(".yml"), true);
-  assert(formats.get(".yml")?.removeExtension);
-  assert(formats.get(".yml")?.page);
+  equals(formats.get(".yml")?.page, "html");
 
   const loader = () => Promise.resolve({});
   const engine = {
@@ -180,12 +173,10 @@ Deno.test("pages configuration", () => {
 
   equals(formats.size, 12);
   equals(formats.get(".ext1")?.loader, loader);
-  assert(formats.get(".ext1")?.page);
-  assert(formats.get(".ext1")?.removeExtension);
+  equals(formats.get(".ext1")?.page, "html");
   assert(formats.get(".ext1")?.engine);
   equals(formats.get(".ext2")?.loader, loader);
-  assert(formats.get(".ext2")?.page);
-  assert(formats.get(".ext2")?.removeExtension);
+  equals(formats.get(".ext2")?.page, "html");
   assert(formats.get(".ext2")?.engine);
 });
 
@@ -201,12 +192,10 @@ Deno.test("assets configuration", () => {
   equals(formats.size, 11);
   equals(formats.has(".css"), true);
   equals(formats.get(".css")?.loader, loader);
-  assert(formats.get(".css")?.page);
-  assert(!formats.get(".css")?.removeExtension);
+  equals(formats.get(".css")?.page, "asset");
   equals(formats.has(".js"), true);
   equals(formats.get(".js")?.loader, loader);
-  assert(formats.get(".js")?.page);
-  assert(!formats.get(".js")?.removeExtension);
+  equals(formats.get(".js")?.page, "asset");
 });
 
 Deno.test("preprocessor configuration", () => {